stao/requirements.txt
google-cloud-bigquery
google-cloud-secret-manager
google-cloud-storage
pysta==0.0.24
requests
pyproj
shapely
geojson


